\name{cleaned_full_incident_data_1970_mar2025}
\alias{cleaned_full_incident_data_1970_mar2025}
\docType{data}
\title{
Cleaned Gas Pipeline Incident Data from 1970 to March 2025
}
\description{
This dataset contains 605 variables from the combined four datasets incident_gas_distribution_1970_mid1984, incident_gas_distribution_mid1984_feb2004, incident_gas_distribution_mar2004_dec2009, and incident_gas_distribution_jan2010_mar2025, all included and documented separately in this package. This dataset and its variables are the product of the cleaning and joining done in the function merge_data() when most recent data has not been downloaded from the PHMSA website. This dataset was created by the authors of this package.
}
\usage{cleaned_full_incident_data_1970_mar2025}
\format{A dataframe containing 19850 observations of 605 variables.}
\source{
Distribution, Transmission & Gathering, LNG, and Liquid Accident and Incident Data, Pipeline and Hazardous Materials Safety Administration, U.S. Department of Transportation, https://www.phmsa.dot.gov/data-and-statistics/pipeline/distribution-transmission-gathering-lng-and-liquid-accident-and-incident-data
}
\keyword{datasets}
