\name{correct_column_names}
\alias{correct_column_names}
\docType{data}
\title{
Correct Column Names for Most Recent Incident Data
}
\description{
This dataset contains the names of the 468 variables in the most recent gas pipeline incident datasets; any incident dataset from January 2010 onward. These variable names are compared to the variable names in the data in the ZIP file passed into the function load_recent_data() to ensure the file was downloaded from the appropriate link on the PHMSA website.
}
\usage{correct_column_names}
\format{A dataframe containing 468 columns.}
\source{
Distribution, Transmission & Gathering, LNG, and Liquid Accident and Incident Data, Pipeline and Hazardous Materials Safety Administration, U.S. Department of Transportation, https://www.phmsa.dot.gov/data-and-statistics/pipeline/distribution-transmission-gathering-lng-and-liquid-accident-and-incident-data
}
\keyword{datasets}
