\name{incident_gas_distribution_1970_mid1984}
\alias{incident_gas_distribution_1970_mid1984}
\docType{data}
\title{
Gas Pipeline Incident Data from 1970 to Mid 1984
}
\description{
This dataset contains information on gas pipeline incidents occuring from 1970 to mid 1984 in the United States.
}
\usage{incident_gas_distribution_1970_mid1984}
\format{A dataframe containing 14611 observations of 138 variables.}
\source{
Distribution, Transmission & Gathering, LNG, and Liquid Accident and Incident Data, Pipeline and Hazardous Materials Safety Administration, U.S. Department of Transportation, https://www.phmsa.dot.gov/data-and-statistics/pipeline/distribution-transmission-gathering-lng-and-liquid-accident-and-incident-data
}
\keyword{datasets}
