\name{incident_gas_distribution_jan2010_mar2025}
\alias{incident_gas_distribution_jan2010_mar2025}
\docType{data}
\title{
Gas Pipeline Incident Data from January 2010 to March 2025
}
\description{
This dataset contains information on gas pipeline incidents occuring from January 2010 to March 2025 in the United States. A version of this dataset containing the most recent incidents can be downloaded from the PHMSA website.
}
\usage{incident_gas_distribution_jan2010_mar2025}
\format{A dataframe containing 1503 observations of 468 variables.}
\source{
Distribution, Transmission & Gathering, LNG, and Liquid Accident and Incident Data, Pipeline and Hazardous Materials Safety Administration, U.S. Department of Transportation, https://www.phmsa.dot.gov/data-and-statistics/pipeline/distribution-transmission-gathering-lng-and-liquid-accident-and-incident-data
}
\keyword{datasets}
