\name{incident_gas_distribution_mar2004_dec2009}
\alias{incident_gas_distribution_mar2004_dec2009}
\docType{data}
\title{
Gas Pipeline Incident Data from March 2004 to December 2009
}
\description{
This dataset contains information on gas pipeline incidents occuring from March 2004 to December 2009 in the United States.
}
\usage{incident_gas_distribution_mar2004_dec2009}
\format{A dataframe containing 895 observations of 172 variables.}
\source{
Distribution, Transmission & Gathering, LNG, and Liquid Accident and Incident Data, Pipeline and Hazardous Materials Safety Administration, U.S. Department of Transportation, https://www.phmsa.dot.gov/data-and-statistics/pipeline/distribution-transmission-gathering-lng-and-liquid-accident-and-incident-data
}
\keyword{datasets}
