\name{incident_gas_distribution_mid1984_feb2004}
\alias{incident_gas_distribution_mid1984_feb2004}
\docType{data}
\title{
Gas Pipeline Incident Data from Mid 1984 to February 2004
}
\description{
This dataset contains information on gas pipeline incidents occuring from mid 1984 to February 2004 in the United States.
}
\usage{incident_gas_distribution_mid1984_feb2004}
\format{A dataframe containing 2841 observations of 62 variables.}
\source{
Distribution, Transmission & Gathering, LNG, and Liquid Accident and Incident Data, Pipeline and Hazardous Materials Safety Administration, U.S. Department of Transportation, https://www.phmsa.dot.gov/data-and-statistics/pipeline/distribution-transmission-gathering-lng-and-liquid-accident-and-incident-data
}
\keyword{datasets}
