% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_recent_data.R
\name{load_recent_data}
\alias{load_recent_data}
\title{Load Most Recent Gas Pipeline Incident Data from ZIP File}
\usage{
load_recent_data(
  zip_path = NULL,
  name = "incident_gas_distribution_jan2010_present"
)
}
\arguments{
\item{zip_path}{Path to the ZIP file containing the .txt data file.}

\item{name}{Name of the dataframe to create in the global environment. Defaults to 'incident_gas_distribution_jan2010_present'. If name is changed, merge_data() will not use most recent data.}
}
\value{
Invisibly returns NULL. The dataframe is assigned to the global environment.
}
\description{
This function loads the most recent gas pipeline incident data from 2010 to present day from a ZIP file downloaded from the Pipeline and Hazardous Materials Safety Administration's website and assigns it to a dataframe by default called "incident_gas_distribution_jan2010_present" in the user's global environment. If this default name is kept, then merge_data() will merge the previous incident data from 1970 to 2010 with this most recent incident data.
}
\examples{
sample_zip <- system.file("extdata", "incident_sample.zip", package = "pipestats")
load_recent_data(zip_path = sample_zip, name = "sample_data")
print("The most recent incident in the example dataset")
print(paste("occurred on:",as.character(sample_data$LOCAL_DATETIME[[1]])))

}
