% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data}
\alias{merge_data}
\title{Clean and Join Gas Pipeline Incident Datasets from 1970 to Present}
\usage{
merge_data()
}
\value{
A dataframe of the four datasets joined together.
}
\description{
This function prepares the variables in each of the datasets from the four time periods to be joined into one full dataset by renaming direct equivalent variables in df1-3 to their name in df4 and creating equivalent variables to df4 from other variables in df1-3 when they encode equivalent information. This function then joins them, saves it as a .tsv within the package, and returns the data as a dataframe, which the user can then save as a dataframe in their local environment.
}
\examples{
data <- merge_data()
print(paste("From 1970 to the present, there have been",as.character(nrow(data)),"gas pipeline incidents in the United States."))
print(paste("From 1970 to the present, we have",as.character(ncol(data)),"unique variables encoding information on gas pipeline incidents."))

}
