% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_recent_data.R
\name{pull_recent_data}
\alias{pull_recent_data}
\title{Load Most Recent Gas Pipeline Incident Data from ZIP File}
\usage{
pull_recent_data(
  zip_path = NULL,
  name = "incident_gas_distribution_jan2010_present"
)
}
\arguments{
\item{zip_path}{Path to the ZIP file containing the .txt data file.}

\item{name}{Name of the dataframe to create in the global environment. Defaults to 'incident_gas_distribution_jan2010_present'. If name is changed, merge_data() will not use most recent data.}
}
\value{
Invisibly returns NULL. The dataframe is assigned to the global environment.
}
\description{
This function loads the most recent gas pipeline incident data from 2010 to present day from a ZIP file downloaded from the Pipeline and Hazardous Materials Safety Administration's website and assigns it to a dataframe by default called "incident_gas_distribution_jan2010_present" in the user's global environment.
}
\examples{
sample_zip <- system.file("extdata", "incident_sample.zip", package = "pipestats")
pull_recent_data(zip_path = sample_zip, name = "sample_data")
print(paste("The most recent incident in the example dataset occured on:",as.character(sample_data$LOCAL_DATETIME[[1]])))

}
