\name{variable_documentation}
\alias{variable_documentation}
\docType{data}
\title{
Variable Documentation for Cleaned Gas Pipeline Incident Data from 1970 to March 2025
}
\description{
This dataset contains the name, description, notes, and which dataset(s) found in for each of the 605 variables in the cleaned and joined gas pipeline incident data from 1970 to March 2025 created by merge_data(). This documentation was created by the authors of this package. More information on the variables can be found on the PHMSA website.
}
\usage{variable_documentation}
\format{A dataset containing 7 values for 605 variable names.}
\source{
Distribution, Transmission & Gathering, LNG, and Liquid Accident and Incident Data, Pipeline and Hazardous Materials Safety Administration, U.S. Department of Transportation, https://www.phmsa.dot.gov/data-and-statistics/pipeline/distribution-transmission-gathering-lng-and-liquid-accident-and-incident-data
}
\keyword{datasets}
