% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data}
\alias{merge_data}
\title{Clean and Join Gas Pipeline Incident Datasets from 1970 to Present}
\usage{
merge_data()
}
\value{
A dataframe of the four datasets joined together.
}
\description{
This function prepares the variables in each of the datasets from the four time periods to be joined into one full dataset by renaming direct equivalent variables in the first three datasets to their names in the most recent dataset and creating equivalent variables to the most recent dataset's variables from other variables in the first three datasets when they encode equivalent information. This function then joins them and returns the data as a dataframe, which the user can save in their local environment.
}
\examples{
data <- merge_data()
print(paste("Since 1970, there have been",as.character(nrow(data)),"gas pipeline incidents."))

}
