% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_recent_data.R
\name{pull_recent_data}
\alias{pull_recent_data}
\title{Pull Recent Gas Pipeline Incident Data from PHMSA Website}
\usage{
pull_recent_data()
}
\value{
A dataframe of the most recent data
}
\description{
This function pulls the most recent gas pipeline incident data from 2010 to present day from the Pipeline and Hazardous Materials Safety Administration website, saves the data as a .txt within the package, and returns the data as a dataframe, which the user can then save as a dataframe in their local environment.
}
\examples{
data <- pull_recent_data()
print(paste("The most recent incident occurred on:",as.character(data$LOCAL_DATETIME[[1]])))

}
